requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
pdfplumber
PyPDF2
python-multipart
python-dotenv
PyMuPDF
openai
